refactor: tiptap용 커스텀 LocalVideoExtension 최적화 작업
